rename MoviePage to MovieForm
