+++ src/pages/manage/movies/CinemaStep.tsx
@@ -17,7 +17,7 @@ import { Text } from '@/ui/Text';
 import { getValueFromTheme } from '@/ui/theme';
 import { Typeahead } from '@/ui/Typeahead';
 
-import type { FormData } from './MoviePage';
+import type { FormData } from './MovieForm';
 
 const getValue = getValueFromTheme('moviesCreatePage');
