+++ src/pages/manage/movies/NameStep.tsx
@@ -16,7 +16,7 @@ import { Text } from '@/ui/Text';
 import { getValueFromTheme } from '@/ui/theme';
 import { Typeahead } from '@/ui/Typeahead';
 
-import type { FormData } from './MoviePage';
+import type { FormData } from './MovieForm';
 
 type NameStepProps = StackProps & StepProps<FormData>;
