+++ src/pages/manage/movies/[eventId]/edit.page.tsx
@@ -3,7 +3,7 @@ import { dehydrate } from 'react-query/hydration';
 import { useGetEventById } from '@/hooks/api/events';
 import { getApplicationServerSideProps } from '@/utils/getApplicationServerSideProps';
 
-import { MoviePage } from '../MoviePage';
+import { MovieForm } from '../MovieForm';
 
 export const getServerSideProps = getApplicationServerSideProps(
   async ({ req, query, cookies, queryClient }) => {
@@ -24,4 +24,4 @@ export const getServerSideProps = getApplicationServerSideProps(
   },
 );
 
-export default MoviePage;
+export default MovieForm;
